requirements.txt
# DecoPilot 依赖列表

# LangChain 核心
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1

# 向量数据库
chromadb>=0.4.0

# DashScope (阿里云通义千问)
dashscope>=1.14.0

# Web框架
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# 流式处理
streamlit>=1.28.0

# 搜索
duckduckgo-search>=4.0.0

# 环境变量
python-dotenv>=1.0.0

# ============ 新增依赖 ============

# PDF解析
PyPDF2>=3.0.0

# 异步HTTP (爬虫)
aiohttp>=3.9.0

# 网页解析 (爬虫)
beautifulsoup4>=4.12.0

# JWT认证
python-jose[cryptography]>=3.3.0

# API限流
slowapi>=0.1.9
